Вынес функционал Кэша в отдельный статик класс
